requirements.txt
python-dotenv~=1.1.0
neo4j~=5.28.1
fastapi~=0.115.12